requirements.txt
numpy==2.1.1
opencv-python==4.11.0.86
pyserial==3.5
torch==2.6.0
torchvision==0.21.0
ultralytics==8.3.93
ultralytics-thop==2.0.14
pynput==1.8.1